Rename taxonomy sidebar helper
Remove mention of sidebar, this isn't accurate for small screens, also
this follows conventions for related navigation and matches the component
view names.
