+++ lib/govuk_publishing_components/presenters/taxonomy_navigation.rb
@@ -1,19 +1,35 @@
 module GovukPublishingComponents
   module Presenters
-    # @private
-    class TaxonomySidebar
+    class TaxonomyNavigation
+      RELATED_SECTIONS = %w(
+        policies
+        topical_events
+        world_locations
+        statistical_data_sets
+      ).freeze
+
       def initialize(content_item)
         @content_item = ContentItem.new content_item
+        @related_navigation = RelatedNavigationHelper.new(content_item)
       end
 
       def sidebar
         {
           items: related_items,
           collections: collections,
-        }
+        }.merge(related_navigation_links)
       end
 
     private
+
+      def related_navigation_links
+        {}.tap do |hash|
+          @related_navigation.related_navigation.each do |section|
+            key = section.keys.first
+            hash[key.to_sym] = section[key] if RELATED_SECTIONS.include?(key)
+          end
+        end
+      end
 
       def there_are_related_item_overrides?
         # TODO: We should check for any external links when we have "new"
