+++ spec/lib/presenters/taxonomy_navigation_spec.rb
@@ -4,7 +4,7 @@ require 'gds_api/test_helpers/rummager'
 
 include GdsApi::TestHelpers::Rummager
 
-RSpec.describe GovukPublishingComponents::Presenters::TaxonomySidebar do
+RSpec.describe GovukPublishingComponents::Presenters::TaxonomyNavigation do
   describe '#sidebar' do
     it 'can handle any valid content item' do
       stub_any_rummager_search_to_return_no_results
